refactor: reuse tensor-parallel shard gather logic
- move TP shard gathering into parallel utils
- reuse it in TP autograd and LoRA checkpoint paths
- document pending DDP and checkpoint integration work
- simplify LoRA tests with Tensor::To
