+++ infini_train/src/nn/parallel/utils.cc
@@ -1,6 +1,11 @@
 #include "infini_train/include/nn/parallel/utils.h"
 
+#include "glog/logging.h"
+
+#include "infini_train/include/nn/functional.h"
 #include "infini_train/include/nn/parallel/global.h"
+#include "infini_train/include/nn/parallel/process_group.h"
+#include "infini_train/include/tensor.h"
 
 namespace infini_train::nn::parallel {
 
@@ -22,6 +27,37 @@ std::vector<int> GetTensorParallelGroupRanks(int global_rank) { return global::G
 
 std::vector<int> GetPipelineParallelGroupRanks(int global_rank) {
     return global::GetGroupRanks(global::PP, global_rank);
+}
+
+std::shared_ptr<Tensor> GatherTensorParallelShard(const std::shared_ptr<Tensor> &tensor, int64_t dim) {
+    const int tp_size = global::GetTensorParallelSize();
+    CHECK_GT(tp_size, 0) << "Tensor Parallel group not initialized";
+    if (tp_size == 1) {
+        return tensor;
+    }
+
+    if (dim < 0) {
+        dim += static_cast<int64_t>(tensor->Dims().size());
+    }
+    CHECK_GE(dim, 0);
+    CHECK_LT(dim, static_cast<int64_t>(tensor->Dims().size()));
+
+    auto device = tensor->GetDevice();
+    auto *tp_group = ProcessGroupFactory::Instance(device.type())
+                         ->Get(GetTensorParallelProcessGroupName(device.Rank().GlobalRank()));
+
+    std::vector<int64_t> gathered_dims = tensor->Dims();
+    gathered_dims[0] *= tp_size;
+    auto gathered = std::make_shared<Tensor>(gathered_dims, tensor->Dtype(), device);
+    tp_group->AllGather(gathered, tensor, false);
+
+    if (dim == 0) {
+        return gathered;
+    }
+
+    // AllGather stacks shards along dim 0. Restore rank-major shards before concatenating on the requested dimension.
+    auto rank_major_shards = gathered->Split(tensor->Dims()[0], 0);
+    return nn::function::Concat(rank_major_shards, dim)->Contiguous();
 }
 
 } // namespace infini_train::nn::parallel
